backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pydantic==2.9.0
python-multipart==0.0.9
google-genai==1.5.0
python-slugify==8.0.4
python-dotenv==1.0.1
aiofiles==24.1.0
bcrypt==4.2.0
PyJWT==2.9.0
httpx==0.27.0
beautifulsoup4==4.12.3
